+++ scripts/evaluate_single_tactic_of_thm_in_gym_2nd_version.py
@@ -7,20 +7,11 @@ import re
 from typing import List
 from pathlib import Path
 from loguru import logger
+from src.interaction.utils import get_theorem_names_from_code
 
 from src.mwe import Mwe
-from src.trace.trace import TracedTacticState, Tracer
+from src.trace.trace import Tracer
 from src.interaction.gym import Gym, ProofFinished
-
-
-def get_theorem_names_from_code(code: str) -> List[str]:
-    theorem_names = []
-    lines = code.split("\n")
-    for line in lines:
-        match = re.search(r"\btheorem\b\s+(\w+)", line)
-        if match:
-            theorem_names.append(match.group(1))
-    return theorem_names
 
 
 def evaluate_all_tactics_of_file_in_gym(
@@ -47,8 +38,8 @@ def evaluate_all_tactics_of_file_in_gym(
     success = 0
     for theorem_name in theorem_names:
         logger.debug(f"Testing with theorem: {theorem_name}")
-        if theorem_name != "image_coe_closedBall":
-            continue
+        # if theorem_name != "image_coe_closedBall":
+        #     continue
         mwe = Mwe(
             code,
             theorem_name,
