+++ src/interaction/utils.py
@@ -1,4 +1,15 @@
+import re
 from typing import List
+
+
+def get_theorem_names_from_code(code: str) -> List[str]:
+    theorem_names = []
+    lines = code.split("\n")
+    for line in lines:
+        match = re.search(r"\btheorem\b\s+(\w+)", line)
+        if match:
+            theorem_names.append(match.group(1))
+    return theorem_names
 
 
 def format_tactic_for_repl(proof: str) -> List[str]:
