+++ tests/e2e/test_extracted_tactics_from_func.py
@@ -40,6 +40,6 @@ def test_extracted_theorem_interaction_in_gym() -> None:
     with Gym(mwe) as (gym, state_0):
         state_1 = gym.run_tacs(
             state_0,
-            [(code[tactic.pos : tactic.end_pos]).decode("utf-8") for tactic in tactics],
+            [tactic.get_syntax_of_tactic(code) for tactic in tactics],
         )
         assert isinstance(state_1, ProofFinished)
